api/requirements.txt
flask
flask-cors
anthropic
openai
python-dotenv
modal-client
